Worked on Sign Up a little more
Had a few problems saving data with ParseUser, for some odd reason it
does not want to communicate with it. I removed automatic user making
from the on create and it still does not create a Parse User. If i
transition it to a ParseObject then it works but not ParseUser