Added ticket number section to sign up
I figured this was needed before assigning numbers to people. Its still
not working perfectly as i cant cache data from the enclosed statement
but i will work on it. I really miss lydia she is leaving tomorrow and
its going to suck alot and im going to be moppy.